refactor(cli): split cli.py into sibling modules to enforce 2k line limit
Splits abicheck/cli.py (3155 lines) into four modules so no file exceeds
the AI-readiness hard limit of 2000 lines, then removes the allowlist
escape hatch from scripts/check_ai_readiness.py so the gate is enforced
without exceptions.

Splits:
- abicheck/cli.py: 1949 lines (was 3155) — keeps Click group, shared
  helpers, dump / compare / appcompat / suggest-suppressions / deps /
  stack-check commands. Imports the three sub-modules below at module
  tail for side-effect command registration.
- abicheck/cli_compare_release.py: compare-release command + its 12
  helpers + _RELEASE_VERDICT_ORDER.
- abicheck/cli_baseline.py: baseline push / pull / list / delete group.
- abicheck/cli_debian_symbols.py: debian-symbols generate / validate /
  diff group.

Each sub-module imports `main` and the helpers it needs back from
abicheck.cli. The cycle is intentional (resolved at end of cli.py
load) and now allow-listed in IMPORT_CYCLE_ALLOWLIST inside the
AI-readiness check, with a docstring explaining the pattern.

Adds the new modules to the mypy `disallow_untyped_decorators = false`
override in pyproject.toml so Click decorators stay clean. Mypy stays
at the documented 27-error baseline.

Removes LARGE_FILE_ALLOWLIST from scripts/check_ai_readiness.py — the
2000-line ERROR threshold now applies to every source file. The four
files at 1500-1633 lines (cli.py, dumper.py, diff_platform.py,
compat/cli.py) still WARN as a refactor signal.

Updates CLAUDE.md with the new layout and a step-by-step guide for
adding a new top-level command.

All 6020 fast tests pass; ruff check clean on new files.
