+++ abicheck/cli_compare_release.py
@@ -1,0 +1,953 @@
+# Copyright 2026 Nikolay Petrov
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""CLI — compare-release command and its helpers.
+
+Split out of :mod:`abicheck.cli` to keep that module under the
+AI-readiness file-size limit. Imported for side-effect at the bottom
+of :mod:`abicheck.cli` so the ``@main.command("compare-release")``
+decorator runs.
+"""
+from __future__ import annotations
+
+import json
+import sys
+import tempfile
+from collections.abc import Callable
+from pathlib import Path
+
+import click
+
+from .bundle import BundleDiffResult
+from .checker import DiffResult
+from .cli import (
+    _build_match_map,
+    _collect_release_inputs,
+    _load_suppression_and_policy,
+    _run_compare_pair,
+    _safe_write_output,
+    _setup_verbosity,
+    _write_or_echo,
+    _write_release_step_summary,
+    main,
+)
+from .model import AbiSnapshot
+from .reporter import to_json
+
+# ---------------------------------------------------------------------------
+# compare-release helpers
+# ---------------------------------------------------------------------------
+
+_RELEASE_VERDICT_ORDER: dict[str, int] = {
+    "NO_CHANGE": 0, "COMPATIBLE": 1, "COMPATIBLE_WITH_RISK": 2,
+    "API_BREAK": 3, "BREAKING": 4, "ERROR": 5,
+}
+
+
+_CompareReleaseCommonArgs = tuple[
+    dict[str, Path], dict[str, Path],
+    Path | None, Path | None,
+    Callable[[Path, Path], Path | None],
+    list[Path], list[Path],
+    list[Path], list[Path],
+    str, str,
+    str, Path | None,
+    str, Path | None,
+    Path | None,
+]
+
+
+def _discover_files(
+    input_dir: Path, lib_dir: Path,
+    include_private: bool,
+    discover_shared_libraries: Callable[..., list[Path]], is_package: Callable[[Path], bool],
+) -> list[Path]:
+    """Discover library files from a directory or extracted package."""
+    if is_package(input_dir):
+        files = discover_shared_libraries(lib_dir, include_private=include_private)
+        if not files:
+            files = _collect_release_inputs(lib_dir)
+    else:
+        files = _collect_release_inputs(lib_dir)
+    return files
+
+
+def _resolve_release_headers(
+    headers: tuple[Path, ...],
+    old_headers_only: tuple[Path, ...],
+    new_headers_only: tuple[Path, ...],
+    old_header_dir: Path | None,
+    new_header_dir: Path | None,
+) -> tuple[list[Path], list[Path]]:
+    """Resolve per-side headers for compare-release."""
+    old_h: list[Path] = list(old_headers_only) if old_headers_only else list(headers)
+    new_h: list[Path] = list(new_headers_only) if new_headers_only else list(headers)
+    if old_header_dir and not old_headers_only:
+        old_h = [old_header_dir]
+    if new_header_dir and not new_headers_only:
+        new_h = [new_header_dir]
+    return old_h, new_h
+
+
+def _match_release_keys(
+    old_dir: Path, new_dir: Path,
+    old_map: dict[str, Path], new_map: dict[str, Path],
+    old_files: list[Path], new_files: list[Path],
+    is_package: Callable[[Path], bool],
+) -> tuple[list[str], list[str], list[str], dict[str, Path], dict[str, Path]]:
+    """Match library keys between old and new, handling direct file pairs."""
+    direct_file_pair = (
+        old_dir.is_file() and new_dir.is_file()
+        and not is_package(old_dir) and not is_package(new_dir)
+    )
+    if direct_file_pair:
+        matched_keys = ["__direct_pair__"]
+        old_map = {"__direct_pair__": old_files[0]}
+        new_map = {"__direct_pair__": new_files[0]}
+        return matched_keys, [], [], old_map, new_map
+
+    matched_keys = sorted(set(old_map) & set(new_map))
+    removed_keys = sorted(set(old_map) - set(new_map))
+    added_keys = sorted(set(new_map) - set(old_map))
+    return matched_keys, removed_keys, added_keys, old_map, new_map
+
+
+def _collect_release_warnings(
+    warning_msgs: list[str],
+    matched_keys: list[str], removed_keys: list[str], added_keys: list[str],
+    old_map: dict[str, Path], new_map: dict[str, Path],
+) -> None:
+    """Collect warning messages for unmatched libraries."""
+    for k in removed_keys:
+        warning_msgs.append(f"Warning: library removed: {old_map[k].name}")
+    for k in added_keys:
+        warning_msgs.append(f"Info: library added: {new_map[k].name}")
+    if not matched_keys:
+        warning_msgs.append(
+            "Warning: no matching library pairs found between OLD and NEW inputs."
+        )
+
+
+def _compare_one_library(
+    key: str,
+    old_map: dict[str, Path], new_map: dict[str, Path],
+    old_debug_dir: Path | None, new_debug_dir: Path | None,
+    resolve_debug_info: Callable[[Path, Path], Path | None],
+    old_h: list[Path], new_h: list[Path],
+    old_inc: list[Path], new_inc: list[Path],
+    old_version: str, new_version: str,
+    lang: str, suppress: Path | None,
+    policy: str, policy_file_path: Path | None,
+    output_dir: Path | None,
+) -> dict[str, object]:
+    """Compare one library pair — suitable for parallel dispatch.
+
+    The entire per-library flow (debug info resolution, comparison, output
+    writing) is wrapped so that *any* exception yields an ERROR entry
+    instead of aborting the whole release comparison.
+
+    The full :class:`DiffResult` is stashed in the returned dict under
+    the ``"_diff_result"`` key. Callers that need the full diff (the
+    bundle layer, JUnit aggregation) pop it from the entry before
+    JSON-serialising — keeps the per-library compare a single-pass.
+    """
+    old_path = old_map[key]
+    new_path = new_map[key]
+    try:
+        old_dbg = resolve_debug_info(old_path, old_debug_dir) if old_debug_dir else None
+        new_dbg = resolve_debug_info(new_path, new_debug_dir) if new_debug_dir else None
+        result, _, _ = _run_compare_pair(
+            old_path, new_path,
+            old_h, new_h, old_inc, new_inc,
+            old_version, new_version,
+            lang, suppress, policy, policy_file_path,
+            old_pdb_path=old_dbg, new_pdb_path=new_dbg,
+        )
+        v = result.verdict.value
+        entry: dict[str, object] = {
+            "library": old_path.name, "verdict": v,
+            "breaking": len(result.breaking),
+            "source_breaks": len(result.source_breaks),
+            "risk_changes": len(result.risk),
+            "compatible_additions": len(result.compatible),
+            "_diff_result": result,
+        }
+        if output_dir:
+            lib_report_path = output_dir / f"{old_path.stem}.json"
+            _safe_write_output(lib_report_path, to_json(result))
+        return entry
+    except (click.ClickException, click.UsageError) as exc:
+        return {"library": old_path.name, "verdict": "ERROR", "error": exc.format_message()}
+    except Exception as exc:
+        return {"library": old_path.name, "verdict": "ERROR", "error": str(exc)}
+
+
+def _compare_release_libraries(
+    matched_keys: list[str],
+    old_map: dict[str, Path], new_map: dict[str, Path],
+    old_debug_dir: Path | None, new_debug_dir: Path | None,
+    resolve_debug_info: Callable[[Path, Path], Path | None],
+    old_h: list[Path], new_h: list[Path],
+    old_inc: list[Path], new_inc: list[Path],
+    old_version: str, new_version: str,
+    lang: str, suppress: Path | None,
+    policy: str, policy_file_path: Path | None,
+    output_dir: Path | None,
+    collect_diff_results: bool = False,
+    *,
+    annotate: bool = False,
+    annotate_additions: bool = False,
+    jobs: int = 1,
+) -> tuple[list[dict[str, object]], str, list[tuple[DiffResult, AbiSnapshot]]]:
+    """Compare each matched library pair and collect results.
+
+    When *collect_diff_results* is True, ``(DiffResult, old_snapshot)``
+    pairs are collected and returned as the third element of the tuple
+    (used by the JUnit output format).
+
+    When *jobs* > 1, comparisons are dispatched in parallel via
+    :func:`_compare_one_library` using a :class:`ProcessPoolExecutor`.
+    """
+    import os as _os
+
+    effective_jobs = jobs if jobs > 0 else (_os.cpu_count() or 1)
+    library_results: list[dict[str, object]] = []
+    diff_pairs: list[tuple[DiffResult, AbiSnapshot]] = []
+    worst_verdict = "NO_CHANGE"
+    all_annotations: list[tuple[int, str]] = []
+
+    common_args = (
+        old_map, new_map, old_debug_dir, new_debug_dir, resolve_debug_info,
+        old_h, new_h, old_inc, new_inc,
+        old_version, new_version,
+        lang, suppress, policy, policy_file_path, output_dir,
+    )
+
+    if effective_jobs > 1 and len(matched_keys) > 1:
+        library_results.extend(
+            _compare_release_parallel(matched_keys, common_args, old_map, effective_jobs),
+        )
+    else:
+        library_results.extend(
+            _compare_release_sequential(matched_keys, common_args),
+        )
+
+    # Post-process all results: compute worst verdict, collect annotations,
+    # and optionally collect diff_pairs (for JUnit).
+    for entry in library_results:
+        v = str(entry["verdict"])
+        if v == "ERROR":
+            if "error" in entry:
+                click.echo(f"Error comparing {entry['library']}: {entry['error']}", err=True)
+        if _RELEASE_VERDICT_ORDER.get(v, 0) > _RELEASE_VERDICT_ORDER.get(worst_verdict, 0):
+            worst_verdict = v
+
+    # collect_diff_results and annotate require re-running comparison for
+    # affected libraries (only used for JUnit / GitHub annotations which
+    # are sequential-only features)
+    if collect_diff_results or annotate:
+        extra_pairs, extra_annotations = _collect_release_extras(
+            matched_keys, old_map, new_map,
+            old_debug_dir, new_debug_dir, resolve_debug_info,
+            old_h, new_h, old_inc, new_inc,
+            old_version, new_version, lang,
+            suppress, policy, policy_file_path,
+            annotate_additions=annotate_additions,
+            collect_diff_results=collect_diff_results,
+            annotate=annotate,
+        )
+        diff_pairs.extend(extra_pairs)
+        all_annotations.extend(extra_annotations)
+
+    # Emit annotations once: sort globally across all libraries by severity,
+    # then truncate to the cap.  This ensures the most important annotations
+    # (errors) are always visible regardless of which library they came from.
+    if all_annotations:
+        from .annotations import format_annotations
+
+        text = format_annotations(all_annotations)
+        if text:
+            click.echo(text, err=True)
+
+    return library_results, worst_verdict, diff_pairs
+
+
+def _compare_release_parallel(
+    matched_keys: list[str],
+    common_args: _CompareReleaseCommonArgs,
+    old_map: dict[str, Path],
+    max_workers: int,
+) -> list[dict[str, object]]:
+    """Run per-library release comparisons in parallel."""
+    from concurrent.futures import ThreadPoolExecutor, as_completed
+
+    results: list[dict[str, object]] = []
+    with ThreadPoolExecutor(max_workers=max_workers) as executor:
+        futures = {
+            executor.submit(_compare_one_library, key, *common_args): key
+            for key in matched_keys
+        }
+        for future in as_completed(futures):
+            key = futures[future]
+            try:
+                results.append(future.result())
+            except Exception as exc:
+                click.echo(f"Error comparing {old_map[key].name}: {exc}", err=True)
+                results.append(
+                    {"library": old_map[key].name, "verdict": "ERROR", "error": str(exc)},
+                )
+    return results
+
+
+def _compare_release_sequential(
+    matched_keys: list[str],
+    common_args: _CompareReleaseCommonArgs,
+) -> list[dict[str, object]]:
+    """Run per-library release comparisons sequentially."""
+    return [_compare_one_library(key, *common_args) for key in matched_keys]
+
+
+def _collect_release_extras(
+    matched_keys: list[str],
+    old_map: dict[str, Path],
+    new_map: dict[str, Path],
+    old_debug_dir: Path | None,
+    new_debug_dir: Path | None,
+    resolve_debug_info: Callable[[Path, Path], Path | None],
+    old_h: list[Path],
+    new_h: list[Path],
+    old_inc: list[Path],
+    new_inc: list[Path],
+    old_version: str,
+    new_version: str,
+    lang: str,
+    suppress: Path | None,
+    policy: str,
+    policy_file_path: Path | None,
+    *,
+    annotate_additions: bool,
+    collect_diff_results: bool,
+    annotate: bool,
+) -> tuple[list[tuple[DiffResult, AbiSnapshot]], list[tuple[int, str]]]:
+    """Collect optional re-run artifacts for JUnit and annotations."""
+    diff_pairs: list[tuple[DiffResult, AbiSnapshot]] = []
+    annotations: list[tuple[int, str]] = []
+    for key in matched_keys:
+        old_path = old_map[key]
+        new_path = new_map[key]
+        old_dbg = resolve_debug_info(old_path, old_debug_dir) if old_debug_dir else None
+        new_dbg = resolve_debug_info(new_path, new_debug_dir) if new_debug_dir else None
+        try:
+            result, old_snap, _ = _run_compare_pair(
+                old_path, new_path,
+                old_h, new_h, old_inc, new_inc,
+                old_version, new_version,
+                lang, suppress, policy, policy_file_path,
+                old_pdb_path=old_dbg, new_pdb_path=new_dbg,
+            )
+        except Exception as exc:
+            click.echo(
+                f"Warning: failed to re-run comparison for {old_path.name}: {exc}",
+                err=True,
+            )
+            continue
+        if collect_diff_results:
+            diff_pairs.append((result, old_snap))
+        if annotate:
+            from .annotations import collect_annotations, is_github_actions
+
+            if is_github_actions():
+                annotations.extend(
+                    collect_annotations(result, annotate_additions=annotate_additions),
+                )
+    return diff_pairs, annotations
+
+
+def _run_bundle_analysis(
+    old_map: dict[str, Path],
+    new_map: dict[str, Path],
+    per_lib_results: list[DiffResult],
+    *,
+    manifest_path: Path | None,
+    bundle_system_providers: str,
+) -> BundleDiffResult | None:
+    """Run bundle-level (ADR-023) analysis on a compare-release run.
+
+    Reuses the per-library :class:`DiffResult`s already computed by
+    :func:`_compare_release_libraries` — no second per-pair compare pass.
+
+    Returns None when there is nothing to analyze (e.g. all libraries
+    failed to dump). Errors during analysis are caught and reported as a
+    warning rather than aborting; bundle analysis is additive.
+    """
+    from .bundle import (
+        BundleDiffResult,
+        build_bundle_snapshot,
+        compare_bundle,
+        load_manifest,
+    )
+
+    if not old_map and not new_map:
+        return None
+    try:
+        old_snap = build_bundle_snapshot(dict(old_map))
+        new_snap = build_bundle_snapshot(dict(new_map))
+    except Exception as exc:
+        # Treat snapshot-build failures as additive degradation: the
+        # per-library compare-release report is still useful, and the
+        # user has an obvious escape hatch (--no-bundle-analysis) if they
+        # want to silence this. A surprise CLI exit here would block CI
+        # pipelines that previously didn't see bundle analysis at all.
+        click.echo(f"Warning: bundle analysis skipped: {exc}", err=True)
+        return None
+
+    manifest = None
+    if manifest_path is not None:
+        try:
+            manifest = load_manifest(manifest_path)
+        except Exception as exc:
+            # Manifest is an *explicit* user input. A malformed --manifest
+            # is a user error, not an environmental quirk; fail loudly so
+            # the contract violation isn't hidden behind a stderr warning.
+            raise click.ClickException(
+                f"Failed to load manifest {manifest_path}: {exc}",
+            ) from exc
+
+    system_extra: list[str] = [
+        s.strip() for s in bundle_system_providers.split(",") if s.strip()
+    ]
+    try:
+        return compare_bundle(
+            old_snap, new_snap, per_lib_results,
+            manifest=manifest,
+            system_providers=system_extra or None,
+        )
+    except Exception as exc:
+        # Analysis-engine bugs should not block the per-library report;
+        # surface as a warning. Future work: surface as a coverage_warning
+        # in the JSON output so downstream CI can detect degradation.
+        click.echo(f"Warning: bundle analysis raised: {exc}", err=True)
+        return BundleDiffResult(old_root=old_snap.root, new_root=new_snap.root)
+
+
+def _format_release_summary(
+    fmt: str, worst_verdict: str,
+    old_dir: Path, new_dir: Path,
+    library_results: list[dict[str, object]],
+    removed_keys: list[str], added_keys: list[str],
+    old_map: dict[str, Path], new_map: dict[str, Path],
+    warning_msgs: list[str],
+    diff_pairs: list[tuple[DiffResult, AbiSnapshot]] | None = None,
+    bundle_result: BundleDiffResult | None = None,
+) -> str:
+    """Format the release comparison summary as JSON, markdown, or JUnit XML."""
+    if fmt == "junit":
+        from .junit_report import to_junit_xml_multi
+        pairs: list[tuple[DiffResult, AbiSnapshot | None]] = list(diff_pairs or [])
+        error_libs = [
+            entry for entry in library_results
+            if entry.get("verdict") == "ERROR"
+        ]
+        return to_junit_xml_multi(
+            pairs, error_libraries=error_libs if error_libs else None,
+        )
+
+    if fmt == "json":
+        summary: dict[str, object] = {
+            "verdict": worst_verdict,
+            "old_dir": str(old_dir),
+            "new_dir": str(new_dir),
+            "libraries": library_results,
+            "unmatched_old": [old_map[k].name for k in removed_keys],
+            "unmatched_new": [new_map[k].name for k in added_keys],
+            "warnings": warning_msgs,
+        }
+        if bundle_result is not None:
+            summary["bundle_verdict"] = bundle_result.bundle_verdict.value
+            summary["bundle_findings"] = [
+                {
+                    "kind": f.kind.value,
+                    "symbol": f.symbol,
+                    "consumer_library": f.consumer_library,
+                    "provider_library": f.provider_library,
+                    "description": f.description,
+                    "old_value": f.old_value,
+                    "new_value": f.new_value,
+                    "affected_libraries": list(f.affected_libraries),
+                }
+                for f in bundle_result.bundle_findings
+            ]
+        return json.dumps(summary, indent=2)
+
+    _VERDICT_EMOJI = {
+        "NO_CHANGE": "✅", "COMPATIBLE": "✅", "COMPATIBLE_WITH_RISK": "⚠️",
+        "API_BREAK": "⚠️", "BREAKING": "❌", "ERROR": "💥",
+    }
+    lines: list[str] = [
+        "# ABI Release Comparison", "",
+        "| | |", "|---|---|",
+        f"| **Old** | `{old_dir}` |",
+        f"| **New** | `{new_dir}` |",
+        f"| **Verdict** | {_VERDICT_EMOJI.get(worst_verdict, '?')} `{worst_verdict}` |",
+    ]
+    bundle_count = len(bundle_result.bundle_findings) if bundle_result else 0
+    if bundle_result is not None:
+        bundle_em = _VERDICT_EMOJI.get(bundle_result.bundle_verdict.value, "?")
+        lines.append(
+            f"| **Bundle** | {bundle_em} `{bundle_result.bundle_verdict.value}` "
+            f"({bundle_count} cross-library finding{'s' if bundle_count != 1 else ''}) |",
+        )
+    lines += [
+        "", "## Libraries", "",
+        "| Library | Verdict | Breaking | Source | Risk | Additions |",
+        "|---|---|---|---|---|---|",
+    ]
+    for lib in library_results:
+        em = _VERDICT_EMOJI.get(str(lib["verdict"]), "?")
+        lines.append(
+            f"| `{lib['library']}` | {em} `{lib['verdict']}` "
+            f"| {lib.get('breaking', '—')} | {lib.get('source_breaks', '—')} "
+            f"| {lib.get('risk_changes', '—')} | {lib.get('compatible_additions', '—')} |"
+        )
+    if removed_keys:
+        lines += ["", "## ⚠️ Removed Libraries", ""]
+        for k in removed_keys:
+            lines.append(f"- `{old_map[k].name}`")
+    if added_keys:
+        lines += ["", "## ℹ️ Added Libraries", ""]
+        for k in added_keys:
+            lines.append(f"- `{new_map[k].name}`")
+    if bundle_result is not None and bundle_result.bundle_findings:
+        lines += ["", "## 🔗 Bundle (Cross-Library) Findings", ""]
+        for f in bundle_result.bundle_findings:
+            # Library-scoped findings (bundle_library_added /
+            # bundle_library_removed) carry the library name in
+            # `symbol`; manifest/import findings carry the symbol.
+            # Both are non-empty in practice, but guard against future
+            # finding shapes with no symbol attribution.
+            lines.append(
+                f"- **{f.kind.value}**"
+                + (f" — `{f.symbol}`" if f.symbol else "")
+                + (f" (consumer: `{f.consumer_library}`)" if f.consumer_library else "")
+                + (f" (provider: `{f.provider_library}`)" if f.provider_library else ""),
+            )
+            lines.append(f"  - {f.description}")
+    return "\n".join(lines)
+
+
+def _write_release_summary_file(
+    output_dir: Path, worst_verdict: str,
+    library_results: list[dict[str, object]],
+    removed_keys: list[str], added_keys: list[str],
+    old_map: dict[str, Path], new_map: dict[str, Path],
+) -> None:
+    """Write per-library summary JSON to output directory."""
+    summary_data: dict[str, object] = {
+        "verdict": worst_verdict,
+        "libraries": library_results,
+        "unmatched_old": [old_map[k].name for k in removed_keys],
+        "unmatched_new": [new_map[k].name for k in added_keys],
+    }
+    summary_path = output_dir / "summary.json"
+    _safe_write_output(summary_path, json.dumps(summary_data, indent=2))
+    click.echo(f"Per-library reports written to {output_dir}/", err=True)
+
+
+@main.command("compare-release")
+@click.argument("old_dir", type=click.Path(exists=True, path_type=Path))
+@click.argument("new_dir", type=click.Path(exists=True, path_type=Path))
+@click.option("-H", "--header", "headers", multiple=True,
+              type=click.Path(path_type=Path),
+              help="Public header file or directory applied to both sides.")
+@click.option("-I", "--include", "includes", multiple=True,
+              type=click.Path(path_type=Path),
+              help="Extra include directory for castxml.")
+@click.option("--old-header", "old_headers_only", multiple=True,
+              type=click.Path(path_type=Path),
+              help="Header for old side only (overrides -H for old).")
+@click.option("--new-header", "new_headers_only", multiple=True,
+              type=click.Path(path_type=Path),
+              help="Header for new side only (overrides -H for new).")
+@click.option("--old-version", "old_version", default="old", show_default=True,
+              help="Version label for old side.")
+@click.option("--new-version", "new_version", default="new", show_default=True,
+              help="Version label for new side.")
+@click.option("--lang", default="c++", show_default=True,
+              type=click.Choice(["c++", "c"], case_sensitive=False))
+@click.option("--format", "fmt",
+              type=click.Choice(["json", "markdown", "junit"]),
+              default="markdown", show_default=True)
+@click.option("-o", "--output", type=click.Path(path_type=Path), default=None,
+              help="Output file for summary report (default: stdout).")
+@click.option("--output-dir", "output_dir", type=click.Path(path_type=Path), default=None,
+              help="Directory to write per-library reports.")
+@click.option("--suppress", type=click.Path(exists=True, path_type=Path), default=None,
+              help="Suppression file (YAML).")
+@click.option("--strict-suppressions", is_flag=True, default=False,
+              help="Fail with exit code 1 if any suppression rule has expired.")
+@click.option("--require-justification", is_flag=True, default=False,
+              help="Require every suppression rule to have a non-empty 'reason' field.")
+@click.option("--policy", "policy",
+              type=click.Choice(["strict_abi", "sdk_vendor", "plugin_abi"], case_sensitive=True),
+              default="strict_abi", show_default=True)
+@click.option("--policy-file", "policy_file_path",
+              type=click.Path(exists=True, path_type=Path), default=None)
+@click.option("--fail-on-removed-library/--no-fail-on-removed-library",
+              "fail_on_removed", default=False,
+              help="Exit 8 when a library present in old_dir is absent in new_dir.")
+@click.option("--debug-info1", type=click.Path(exists=True, path_type=Path), default=None,
+              help="Debug info package for old side (RPM/Deb/tar).")
+@click.option("--debug-info2", type=click.Path(exists=True, path_type=Path), default=None,
+              help="Debug info package for new side (RPM/Deb/tar).")
+@click.option("--devel-pkg1", type=click.Path(exists=True, path_type=Path), default=None,
+              help="Development package with headers for old side.")
+@click.option("--devel-pkg2", type=click.Path(exists=True, path_type=Path), default=None,
+              help="Development package with headers for new side.")
+@click.option("--dso-only", is_flag=True, default=False,
+              help="Only compare shared objects, skip executables.")
+@click.option("--include-private-dso", is_flag=True, default=False,
+              help="Include private (non-public) shared objects from non-standard paths.")
+@click.option("--keep-extracted", is_flag=True, default=False,
+              help="Keep extracted temporary files for debugging.")
+@click.option("--annotate", is_flag=True, default=False,
+              help="Emit GitHub Actions workflow command annotations to stdout. "
+                   "Only effective when GITHUB_ACTIONS=true.")
+@click.option("--annotate-additions", is_flag=True, default=False,
+              help="Include additions/compatible changes as ::notice annotations "
+                   "(requires --annotate).")
+@click.option("-v", "--verbose", is_flag=True, default=False)
+@click.option("-j", "--jobs", "jobs", type=int, default=1, show_default=True,
+              help="Number of parallel library comparisons. "
+                   "Use 0 for auto-detect (CPU count).")
+@click.option("--manifest", "manifest_path", type=click.Path(exists=True, path_type=Path),
+              default=None,
+              help="ABI instantiation manifest (YAML/JSON) listing symbols the "
+                   "release publicly promises. See ADR-023.")
+@click.option("--bundle-system-providers", "bundle_system_providers", default="",
+              help="Comma-separated extra sonames to treat as system-provided "
+                   "(extends the built-in libc/libstdc++/libgcc/libtbb allow-list).")
+@click.option("--no-bundle-analysis", "no_bundle_analysis", is_flag=True, default=False,
+              help="Skip bundle-level cross-library analysis (debug/parity escape hatch). "
+                   "Bundle findings catch intra-bundle symbol removals, signature drift "
+                   "across DSO boundaries, type drift across siblings, provider "
+                   "migration, and manifest mismatches.")
+def compare_release_cmd(
+    old_dir: Path,
+    new_dir: Path,
+    headers: tuple[Path, ...],
+    includes: tuple[Path, ...],
+    old_headers_only: tuple[Path, ...],
+    new_headers_only: tuple[Path, ...],
+    old_version: str,
+    new_version: str,
+    lang: str,
+    fmt: str,
+    output: Path | None,
+    output_dir: Path | None,
+    suppress: Path | None,
+    strict_suppressions: bool,
+    require_justification: bool,
+    policy: str,
+    policy_file_path: Path | None,
+    fail_on_removed: bool,
+    debug_info1: Path | None,
+    debug_info2: Path | None,
+    devel_pkg1: Path | None,
+    devel_pkg2: Path | None,
+    dso_only: bool,
+    include_private_dso: bool,
+    keep_extracted: bool,
+    annotate: bool,
+    annotate_additions: bool,
+    verbose: bool,
+    jobs: int,
+    manifest_path: Path | None,
+    bundle_system_providers: str,
+    no_bundle_analysis: bool,
+) -> None:
+    """Compare all libraries in two release directories or packages.
+
+    OLD_DIR and NEW_DIR may each be a file, directory, or package
+    (RPM, Deb, tar, conda, wheel). Package format is auto-detected.
+    When directories are given, libraries are matched by filename stem.
+
+    \b
+    Exit codes:
+      0  All libraries: NO_CHANGE, COMPATIBLE, or COMPATIBLE_WITH_RISK
+      2  At least one library: API_BREAK
+      4  At least one library: BREAKING
+      8  Library removed (only when --fail-on-removed-library)
+
+    \b
+    Examples:
+      abicheck compare-release release-1.0/ release-2.0/ -H include/
+      abicheck compare-release libfoo-1.0.rpm libfoo-1.1.rpm
+      abicheck compare-release libfoo_1.0.deb libfoo_1.1.deb
+      abicheck compare-release sdk-2.0.tar.gz sdk-2.1.tar.gz
+      abicheck compare-release pkg-v1.conda pkg-v2.conda
+      abicheck compare-release old.whl new.whl
+      abicheck compare-release libfoo-1.0.rpm libfoo-1.1.rpm \\
+          --debug-info1 libfoo-debuginfo-1.0.rpm \\
+          --debug-info2 libfoo-debuginfo-1.1.rpm
+    """
+
+    from .package import (
+        _is_elf_shared_object,
+        detect_extractor,
+        discover_shared_libraries,
+        is_package,
+        resolve_debug_info,
+    )
+
+    _setup_verbosity(verbose)
+
+    if annotate_additions and not annotate:
+        raise click.UsageError("--annotate-additions requires --annotate")
+
+    # Track temporary directory paths for cleanup
+    _temp_dir_paths: list[str] = []
+
+    def _make_temp_dir(prefix: str) -> Path:
+        """Create a temporary directory, tracking it for later cleanup."""
+        path = tempfile.mkdtemp(prefix=prefix)
+        _temp_dir_paths.append(path)
+        return Path(path)
+
+    def _extract_if_package(
+        input_path: Path,
+        debug_pkg: Path | None,
+        devel_pkg: Path | None,
+    ) -> tuple[Path, Path | None, Path | None]:
+        """Extract package to tempdir if needed, return (lib_dir, debug_dir, header_dir)."""
+        if not is_package(input_path):
+            return input_path, None, None
+
+        extractor = detect_extractor(input_path)
+        if extractor is None:
+            raise click.ClickException(f"Unrecognized package format: {input_path}")
+
+        target = _make_temp_dir("abicheck_pkg_")
+
+        result = extractor.extract(input_path, target)
+        lib_dir = result.lib_dir
+        debug_dir = result.debug_dir
+        header_dir = result.header_dir
+
+        # Extract debug info package if provided
+        if debug_pkg is not None:
+            dbg_ext = detect_extractor(debug_pkg)
+            if dbg_ext is None:
+                raise click.ClickException(f"Unrecognized debug package format: {debug_pkg}")
+            dbg_target = _make_temp_dir("abicheck_dbg_")
+            dbg_result = dbg_ext.extract(debug_pkg, dbg_target)
+            debug_dir = dbg_result.lib_dir
+
+        # Extract devel package if provided
+        if devel_pkg is not None:
+            dev_ext = detect_extractor(devel_pkg)
+            if dev_ext is None:
+                raise click.ClickException(f"Unrecognized devel package format: {devel_pkg}")
+            dev_target = _make_temp_dir("abicheck_dev_")
+            dev_result = dev_ext.extract(devel_pkg, dev_target)
+            header_dir = dev_result.lib_dir
+
+        return lib_dir, debug_dir, header_dir
+
+    # Validate suppression file early (before per-library loop)
+    if suppress is not None and (strict_suppressions or require_justification):
+        _load_suppression_and_policy(
+            suppress, policy, policy_file_path,
+            strict_suppressions=strict_suppressions,
+            require_justification=require_justification,
+        )
+
+    try:
+        prep = _prepare_compare_release_inputs(
+            old_dir, new_dir,
+            debug_info1, debug_info2, devel_pkg1, devel_pkg2,
+            include_private_dso, dso_only,
+            headers, old_headers_only, new_headers_only, includes,
+            _extract_if_package, discover_shared_libraries, is_package, _is_elf_shared_object,
+        )
+        (
+            old_debug_dir, new_debug_dir,
+            old_h, new_h, old_inc, new_inc,
+            old_map, new_map, warning_msgs,
+            matched_keys, removed_keys, added_keys,
+        ) = prep
+
+        if fmt != "json":
+            for msg in warning_msgs:
+                click.echo(msg, err=True)
+
+        if output_dir:
+            output_dir.mkdir(parents=True, exist_ok=True)
+
+        bundle_enabled = not no_bundle_analysis
+        # JUnit still re-runs pairs in _collect_release_extras because it
+        # needs old AbiSnapshot too. Bundle analysis reuses the
+        # _diff_result stashed in each library entry from the first pass.
+        library_results, worst_verdict, diff_pairs = _compare_release_libraries(
+            matched_keys, old_map, new_map,
+            old_debug_dir, new_debug_dir, resolve_debug_info,
+            old_h, new_h, old_inc, new_inc,
+            old_version, new_version,
+            lang, suppress, policy, policy_file_path,
+            output_dir,
+            collect_diff_results=(fmt == "junit"),
+            annotate=annotate,
+            annotate_additions=annotate_additions,
+            jobs=jobs,
+        )
+
+        bundle_result = None
+        if bundle_enabled:
+            # Pull the per-library DiffResults out of the entries (stashed
+            # by _compare_one_library) and feed them into the bundle layer
+            # without re-running the per-library compare.
+            stashed_diffs: list[DiffResult] = []
+            for entry in library_results:
+                diff = entry.get("_diff_result") if isinstance(entry, dict) else None
+                if isinstance(diff, DiffResult):
+                    stashed_diffs.append(diff)
+            bundle_result = _run_bundle_analysis(
+                old_map, new_map, stashed_diffs,
+                manifest_path=manifest_path,
+                bundle_system_providers=bundle_system_providers,
+            )
+            if bundle_result is not None:
+                bv = bundle_result.bundle_verdict.value
+                if _RELEASE_VERDICT_ORDER.get(bv, 0) > _RELEASE_VERDICT_ORDER.get(worst_verdict, 0):
+                    worst_verdict = bv
+
+        # Strip _diff_result from entries before they leave this scope —
+        # it's a Python object that can't be JSON-serialised and the
+        # summary formatter doesn't need it.
+        for entry in library_results:
+            if isinstance(entry, dict):
+                entry.pop("_diff_result", None)
+
+        if removed_keys and _RELEASE_VERDICT_ORDER.get(worst_verdict, 0) < _RELEASE_VERDICT_ORDER.get("COMPATIBLE_WITH_RISK", 0):
+            worst_verdict = "COMPATIBLE_WITH_RISK"
+
+        text = _format_release_summary(
+            fmt, worst_verdict, old_dir, new_dir,
+            library_results, removed_keys, added_keys,
+            old_map, new_map, warning_msgs,
+            diff_pairs=diff_pairs if fmt == "junit" else None,
+            bundle_result=bundle_result,
+        )
+        _write_or_echo(output, text)
+
+        # Write a single step summary for the entire release comparison.
+        if annotate:
+            _write_release_step_summary(text, fmt)
+
+        if output_dir:
+            _write_release_summary_file(
+                output_dir, worst_verdict, library_results,
+                removed_keys, added_keys, old_map, new_map,
+            )
+
+        _exit_compare_release(worst_verdict, fail_on_removed, removed_keys)
+    finally:
+        import shutil as _shutil
+        if not keep_extracted:
+            for td_path in _temp_dir_paths:
+                _shutil.rmtree(td_path, ignore_errors=True)
+        elif _temp_dir_paths:
+            kept_paths = ", ".join(_temp_dir_paths)
+            click.echo(
+                f"Extracted files kept in: {kept_paths}",
+                err=True,
+            )
+
+
+def _prepare_compare_release_inputs(
+    old_dir: Path,
+    new_dir: Path,
+    debug_info1: Path | None,
+    debug_info2: Path | None,
+    devel_pkg1: Path | None,
+    devel_pkg2: Path | None,
+    include_private_dso: bool,
+    dso_only: bool,
+    headers: tuple[Path, ...],
+    old_headers_only: tuple[Path, ...],
+    new_headers_only: tuple[Path, ...],
+    includes: tuple[Path, ...],
+    extract_if_package: Callable[[Path, Path | None, Path | None], tuple[Path, Path | None, Path | None]],
+    discover_shared_libraries: Callable[..., list[Path]],
+    is_package: Callable[[Path], bool],
+    is_elf_shared_object: Callable[[Path], bool],
+) -> tuple[
+    Path | None, Path | None,
+    list[Path], list[Path], list[Path], list[Path],
+    dict[str, Path], dict[str, Path], list[str],
+    list[str], list[str], list[str],
+]:
+    """Prepare inputs/maps/keys for compare-release command."""
+    old_lib_dir, old_debug_dir, old_header_dir = extract_if_package(
+        old_dir, debug_info1, devel_pkg1,
+    )
+    new_lib_dir, new_debug_dir, new_header_dir = extract_if_package(
+        new_dir, debug_info2, devel_pkg2,
+    )
+    old_files = _discover_files(
+        old_dir, old_lib_dir, include_private_dso, discover_shared_libraries, is_package,
+    )
+    new_files = _discover_files(
+        new_dir, new_lib_dir, include_private_dso, discover_shared_libraries, is_package,
+    )
+    if dso_only:
+        old_files = [f for f in old_files if is_elf_shared_object(f)]
+        new_files = [f for f in new_files if is_elf_shared_object(f)]
+    old_map, old_warns = _build_match_map(old_files)
+    new_map, new_warns = _build_match_map(new_files)
+    warning_msgs: list[str] = [f"Warning: {warning}" for warning in (old_warns + new_warns)]
+    old_h, new_h = _resolve_release_headers(
+        headers, old_headers_only, new_headers_only, old_header_dir, new_header_dir,
+    )
+    old_inc = list(includes)
+    new_inc = list(includes)
+    matched_keys, removed_keys, added_keys, old_map, new_map = _match_release_keys(
+        old_dir, new_dir, old_map, new_map, old_files, new_files, is_package,
+    )
+    _collect_release_warnings(
+        warning_msgs, matched_keys, removed_keys, added_keys, old_map, new_map,
+    )
+    return (
+        old_debug_dir, new_debug_dir,
+        old_h, new_h, old_inc, new_inc,
+        old_map, new_map, warning_msgs,
+        matched_keys, removed_keys, added_keys,
+    )
+
+
+def _exit_compare_release(
+    worst_verdict: str,
+    fail_on_removed: bool,
+    removed_keys: list[str],
+) -> None:
+    """Exit compare-release with ABI-compatible status code mapping."""
+    if worst_verdict in ("BREAKING", "ERROR"):
+        sys.exit(4)
+    if worst_verdict == "API_BREAK":
+        sys.exit(2)
+    if fail_on_removed and removed_keys:
+        sys.exit(8)
+
+
+# ── Suggest suppressions command ──────────────────────────────────────────────
+
