+++ scripts/check_ai_readiness.py
@@ -49,17 +49,10 @@ SCRIPTS = ROOT / "scripts"
 WARN_LINES = 1500
 ERROR_LINES = 2000
 
-# Files documented in CLAUDE.md ("Files that are large — edit carefully").
-# These are allowed to exceed ERROR_LINES, but a WARN is still emitted so the
-# growth is visible.
-LARGE_FILE_ALLOWLIST: frozenset[str] = frozenset(
-    {
-        "abicheck/cli.py",
-        "abicheck/diff_platform.py",
-        "abicheck/dumper.py",
-        "abicheck/compat/cli.py",
-    }
-)
+# Hard line limit is enforced for every source file. If you find yourself
+# wanting to add an entry, split the file instead — the AI-readiness check is
+# meant to keep modules legible for agents.
+LARGE_FILE_ALLOWLIST: frozenset[str] = frozenset()
 
 # Directories that must contain a CLAUDE.md for per-area agent context.
 REQUIRED_CLAUDE_MD_DIRS: tuple[Path, ...] = (
@@ -442,6 +435,21 @@ def _find_cycles(graph: dict[str, set[str]]) -> list[list[str]]:
     return unique
 
 
+# Intentional import cycles to ignore. Each entry is a frozenset of module
+# short names (no `abicheck.` prefix) that participate in a known, by-design
+# cycle — e.g. Click sub-command modules that register on a parent's group.
+IMPORT_CYCLE_ALLOWLIST: frozenset[frozenset[str]] = frozenset(
+    {
+        # cli.py imports cli_compare_release / cli_baseline / cli_debian_symbols
+        # at module-load tail to register their @main.command(...) decorators;
+        # those sub-modules import `main` and shared helpers back from cli.
+        frozenset({"cli", "cli_compare_release"}),
+        frozenset({"cli", "cli_baseline"}),
+        frozenset({"cli", "cli_debian_symbols"}),
+    }
+)
+
+
 def check_import_cycles(f: Findings) -> None:
     # Build module -> direct abicheck imports.
     all_modules = {_module_name(p) for p in PKG.rglob("*.py")}
@@ -463,6 +471,9 @@ def check_import_cycles(f: Findings) -> None:
 
     cycles = _find_cycles(graph)
     for cyc in cycles:
+        short = frozenset(m.removeprefix("abicheck.") for m in cyc[:-1])
+        if short in IMPORT_CYCLE_ALLOWLIST:
+            continue
         f.err(
             "import-cycles",
             " -> ".join(m.removeprefix("abicheck.") for m in cyc),
